+++ packages/block-editor-tools/src/components/selector/index.jsx
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import {
   useCallback,
   useEffect,
+  useMemo,
   useRef,
   useState,
 } from 'react';
@@ -14,7 +15,7 @@ import { addQueryArgs } from '@wordpress/url';
 import { v4 as uuidv4 } from 'uuid';
 
 // Custom hooks.
-import { useDebounce } from '../../hooks';
+import { useDebounce } from '@wordpress/compose';
 
 // Components.
 import SearchResults from './components/search-results';
@@ -47,13 +48,23 @@ const Selector = ({
   const [isOpen, setIsOpen] = useState(false);
   const [loading, setLoadState] = useState(false);
   const [searchString, setSearchString] = useState('');
+  const [debouncedSearchString, setDebouncedSearchString] = useState('');
   const [selectedItems, setSelectedItems] = useState([]);
 
   // Create ref.
   const ref = useRef();
 
+  // Memoize subType string.
+  const selectedSubTypes = useMemo(() => (subTypes.length > 0 ? subTypes.join(',') : 'any'), [subTypes]);
+
+  const debouncedSetSearchString = useDebounce((value) => {
+    setDebouncedSearchString(value);
+  }, 750);
+
   // Debounce search string from input.
-  const debouncedSearchString = useDebounce(searchString, 750);
+  useEffect(() => {
+    debouncedSetSearchString(searchString);
+  }, [searchString, debouncedSetSearchString]);
 
   /**
    * Make API request for items by search string.
@@ -85,7 +96,7 @@ const Selector = ({
       {
         page,
         search: debouncedSearchString,
-        subtype: subTypes.length > 0 ? subTypes.join(',') : 'any',
+        subtype: selectedSubTypes,
         type,
       },
     );
@@ -115,7 +126,15 @@ const Selector = ({
         }
       })
       .catch((err) => setError(err.message));
-  }, [debouncedSearchString, type, maxPages, multiple, subTypes, selectedItems.length, threshold]);
+  }, [
+    debouncedSearchString,
+    maxPages,
+    multiple,
+    selectedItems.length,
+    selectedSubTypes,
+    threshold,
+    type,
+  ]);
 
   /**
    * On Mount, pre-fill selected buttons, if they exist.