+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @alleyinteractive/tsconfig
 
+## 0.3.0
+
+### Minor Changes
+
+- 88bd605: Security patch: Bump form-data from 3.0.1 to 3.0.4
+
 ## 0.2.0
 
 ### Minor Changes